Move V1 pages to components/v1/
