+++ src/components/v1/V1Project/Pay/index.tsx
@@ -1,8 +1,8 @@
 import { Button, Tooltip } from 'antd'
 import { t, Trans } from '@lingui/macro'
 
-import ConfirmPayOwnerModal from 'components/V1Project/modals/ConfirmPayOwnerModal'
-import PayWarningModal from 'components/V1Project/modals/PayWarningModal'
+import ConfirmPayOwnerModal from 'components/v1/V1Project/modals/ConfirmPayOwnerModal'
+import PayWarningModal from 'components/v1/V1Project/modals/PayWarningModal'
 import InputAccessoryButton from 'components/shared/InputAccessoryButton'
 import FormattedNumberInput from 'components/shared/inputs/FormattedNumberInput'
 import { V1ProjectContext } from 'contexts/v1/projectContext'
@@ -17,7 +17,7 @@ import { decodeFundingCycleMetadata } from 'utils/fundingCycle'
 import { disablePayOverrides } from 'constants/v1/overrides'
 import { readNetwork } from 'constants/networks'
 import { CURRENCY_ETH, CURRENCY_USD } from 'constants/currency'
-import CurrencySymbol from '../../shared/CurrencySymbol'
+import CurrencySymbol from '../../../shared/CurrencySymbol'
 import PayInputSubText from './PayInputSubText'
 import { V1_PROJECT_IDS } from 'constants/v1/projectIds'
